# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BflSettingsResponse": {
        "properties": {
          "api_key_configured": {
            "title": "Api Key Configured",
            "type": "boolean"
          },
          "api_key_masked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Masked"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "calls_enabled": {
            "title": "Calls Enabled",
            "type": "boolean"
          },
          "daily_call_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Call Limit"
          },
          "default_provider": {
            "title": "Default Provider",
            "type": "string"
          },
          "max_estimated_cost_per_job": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Estimated Cost Per Job"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "rate_limit_per_minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Per Minute"
          },
          "restart_required": {
            "title": "Restart Required",
            "type": "boolean"
          },
          "result_path": {
            "title": "Result Path",
            "type": "string"
          },
          "rollout_enabled": {
            "title": "Rollout Enabled",
            "type": "boolean"
          },
          "submit_path": {
            "title": "Submit Path",
            "type": "string"
          },
          "submit_url": {
            "title": "Submit Url",
            "type": "string"
          }
        },
        "required": [
          "api_key_configured",
          "api_key_masked",
          "base_url",
          "calls_enabled",
          "daily_call_limit",
          "default_provider",
          "max_estimated_cost_per_job",
          "model",
          "rate_limit_per_minute",
          "restart_required",
          "result_path",
          "rollout_enabled",
          "submit_path",
          "submit_url"
        ],
        "title": "BflSettingsResponse",
        "type": "object"
      },
      "BflSettingsUpdateRequest": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "default": "https://api.bfl.ai",
            "maxLength": 512,
            "title": "Base Url",
            "type": "string"
          },
          "calls_enabled": {
            "default": false,
            "title": "Calls Enabled",
            "type": "boolean"
          },
          "daily_call_limit": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Call Limit"
          },
          "default_provider": {
            "default": "disabled",
            "maxLength": 80,
            "title": "Default Provider",
            "type": "string"
          },
          "max_estimated_cost_per_job": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Estimated Cost Per Job"
          },
          "model": {
            "default": "flux-2-pro-preview",
            "maxLength": 120,
            "title": "Model",
            "type": "string"
          },
          "rate_limit_per_minute": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Per Minute"
          },
          "result_path": {
            "default": "/v1/get_result",
            "maxLength": 255,
            "title": "Result Path",
            "type": "string"
          },
          "rollout_enabled": {
            "default": false,
            "title": "Rollout Enabled",
            "type": "boolean"
          },
          "submit_path": {
            "default": "/v1/flux-2-pro-preview",
            "maxLength": 255,
            "title": "Submit Path",
            "type": "string"
          }
        },
        "title": "BflSettingsUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpenAISettingsResponse": {
        "properties": {
          "api_key_configured": {
            "title": "Api Key Configured",
            "type": "boolean"
          },
          "api_key_masked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Masked"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "calls_enabled": {
            "title": "Calls Enabled",
            "type": "boolean"
          },
          "daily_call_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Call Limit"
          },
          "default_provider": {
            "title": "Default Provider",
            "type": "string"
          },
          "image_model": {
            "title": "Image Model",
            "type": "string"
          },
          "image_path": {
            "title": "Image Path",
            "type": "string"
          },
          "image_url": {
            "title": "Image Url",
            "type": "string"
          },
          "max_estimated_cost_per_job": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Estimated Cost Per Job"
          },
          "parser_enabled": {
            "title": "Parser Enabled",
            "type": "boolean"
          },
          "rate_limit_per_minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Per Minute"
          },
          "responses_path": {
            "title": "Responses Path",
            "type": "string"
          },
          "restart_required": {
            "title": "Restart Required",
            "type": "boolean"
          },
          "rollout_enabled": {
            "title": "Rollout Enabled",
            "type": "boolean"
          },
          "text_model": {
            "title": "Text Model",
            "type": "string"
          }
        },
        "required": [
          "api_key_configured",
          "api_key_masked",
          "base_url",
          "calls_enabled",
          "daily_call_limit",
          "default_provider",
          "image_model",
          "image_path",
          "image_url",
          "max_estimated_cost_per_job",
          "parser_enabled",
          "rate_limit_per_minute",
          "responses_path",
          "restart_required",
          "rollout_enabled",
          "text_model"
        ],
        "title": "OpenAISettingsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Control-plane API foundation for the carAgent workbench.",
    "title": "carAgent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/operations/bfl-settings": {
      "get": {
        "operationId": "get_bfl_settings_operations_bfl_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BflSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Bfl Settings",
        "tags": [
          "operations"
        ]
      },
      "post": {
        "operationId": "update_bfl_settings_operations_bfl_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BflSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BflSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Bfl Settings",
        "tags": [
          "operations"
        ]
      }
    },
    "/operations/openai-settings": {
      "get": {
        "operationId": "get_openai_settings_operations_openai_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAISettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Openai Settings",
        "tags": [
          "operations"
        ]
      },
      "post": {
        "operationId": "update_openai_settings_operations_openai_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAISettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAISettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Openai Settings",
        "tags": [
          "operations"
        ]
      }
    },
...
  }
}